Move notification code into com.hedvig.android.app
