+++ app/app/src/main/kotlin/com/hedvig/android/app/notification/senders/ReferralsNotificationSender.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.service.push.senders
+package com.hedvig.android.app.notification.senders
 
 import android.app.Notification
 import android.app.PendingIntent
@@ -10,13 +10,13 @@ import androidx.core.app.PendingIntentCompat
 import androidx.core.app.TaskStackBuilder
 import com.google.firebase.messaging.RemoteMessage
 import com.hedvig.android.app.MainActivity
+import com.hedvig.android.app.notification.DATA_MESSAGE_BODY
+import com.hedvig.android.app.notification.DATA_MESSAGE_TITLE
+import com.hedvig.android.app.notification.getImmutablePendingIntentFlags
 import com.hedvig.android.core.common.android.notification.setupNotificationChannel
 import com.hedvig.android.navigation.core.HedvigDeepLinkContainer
 import com.hedvig.android.notification.core.NotificationSender
 import com.hedvig.android.notification.core.sendHedvigNotification
-import com.hedvig.app.service.push.DATA_MESSAGE_BODY
-import com.hedvig.app.service.push.DATA_MESSAGE_TITLE
-import com.hedvig.app.service.push.getImmutablePendingIntentFlags
 import hedvig.resources.R
 
 class ReferralsNotificationSender(
